+++ tests/pyftpdlib-server.py
@@ -13,7 +13,7 @@ import logging
 
 from pyftpdlib.servers import FTPServer
 from pyftpdlib.authorizers import DummyAuthorizer
-from pyftpdlib.handlers import TLS_FTPHandler
+from pyftpdlib.handlers import FTPHandler, TLS_FTPHandler
 
 
 def main(args=None):
@@ -22,6 +22,7 @@ def main(args=None):
     ap.add_argument("-u", "--user", default="joedoe", help="User account name (default: %(default)s)")
     ap.add_argument("-P", "--password", default="abc123", help="User account password (default: %(default)s)")
     ap.add_argument("-c", "--certfile", default="keycert.pem", help="Certificate/Keyfile (default: %(default)s)")
+    ap.add_argument("-s", "--tls", action="store_true", help="Enable/require TLS")
     ap.add_argument("-v", "--verbose", action="store_true", help="Enable verbose (debug) output")
 
     args = ap.parse_args(args)
@@ -32,12 +33,17 @@ def main(args=None):
     authorizer = DummyAuthorizer()
     authorizer.add_user(args.user, args.password, '.', perm='elradfmwMT')
     authorizer.add_anonymous('.')
-    handler = TLS_FTPHandler
-    handler.certfile = args.certfile
+
+    if args.tls:
+        handler = TLS_FTPHandler
+        handler.certfile = args.certfile
+        # requires SSL for both control and data channel
+        handler.tls_control_required = True
+        handler.tls_data_required = True
+    else:
+        handler = FTPHandler
+
     handler.authorizer = authorizer
-    # requires SSL for both control and data channel
-    handler.tls_control_required = True
-    handler.tls_data_required = True
     server = FTPServer(('', args.port), handler)
     server.serve_forever()
