+++ tests/test_ftplibtls.py
@@ -1,4 +1,4 @@
-from ftp_tls import FTP_TLS
+from ftplibtls import FTP_TLS
 
 PORT = 2121
 
@@ -9,6 +9,6 @@ ftp.prot_p()
 files = ftp.nlst()
 print(files)
 
-assert 'test_ftplib.py' in files
+assert 'test_ftplibtly.py' in files
 assert 'pyftpdlib-tls-server.py' in files
 assert 'keycert.pem' in files
